Reorganize examples.
